+++ swiyu-trust-management-service/src/main/java/ch/admin/bj/swiyu/trust/management/modules/common/exception/BusinessPartnerIdentityBadRequestException.java
@@ -1,8 +1,0 @@
-package ch.admin.bj.swiyu.trust.management.modules.common.exception;
-
-public class BusinessPartnerIdentityBadRequestException extends RuntimeException {
-
-    public BusinessPartnerIdentityBadRequestException(String message) {
-        super(message);
-    }
-}