+++ swiyu-trust-management-service/src/main/java/ch/admin/bj/swiyu/trust/management/modules/management/domain/TrustStatementPartnerLinkRepository.java
@@ -71,4 +71,6 @@ public interface TrustStatementPartnerLinkRepository
         UUID protectedIssuanceAuthorizationId,
         TrustStatementPartnerLinkStatus status
     );
+
+    void deleteAllByPartnerId(UUID partnerId);
 }