+++ swiyu-trust-management-service/src/main/java/ch/admin/bj/swiyu/trust/management/modules/management/infrastructure/web/controller/RestExceptionHandler.java
@@ -1,6 +1,6 @@
 package ch.admin.bj.swiyu.trust.management.modules.management.infrastructure.web.controller;
 
-import ch.admin.bj.swiyu.trust.management.modules.common.exception.BusinessPartnerIdentityBadRequestException;
+import ch.admin.bj.swiyu.trust.management.modules.common.exception.BusinessPartnerIdentityNotActiveException;
 import ch.admin.bj.swiyu.trust.management.modules.common.exception.ExternalSystemProxyException;
 import ch.admin.bj.swiyu.trust.management.modules.common.exception.ResourceNotFoundException;
 import ch.admin.bj.swiyu.trust.management.modules.common.exception.TrustStatementValidationFailedException;
@@ -49,10 +49,10 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
         return e.getValidationIssues();
     }
 
-    @ExceptionHandler(BusinessPartnerIdentityBadRequestException.class)
+    @ExceptionHandler(BusinessPartnerIdentityNotActiveException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseEntity<String> handleBusinessPartnerIdentityBadRequestException(
-        BusinessPartnerIdentityBadRequestException e
+        BusinessPartnerIdentityNotActiveException e
     ) {
         log.debug(e.getMessage());
         return new ResponseEntity<>(e.getMessage(), HttpStatusCode.valueOf(400));